Lucas

+++ EPM/src/apresentacao/project/ProjectMenu.java
@@ -24,44 +24,54 @@ import apresentacao.part.PartMenu;
 import negocio.Project;
 
 public class ProjectMenu {
+	//Class Variables
 	
-	private String user = "";
-	//propriedades da classe
+	//Frame and panel
 	private JFrame fProjectMenu = new JFrame();
 	private JPanel pn = new JPanel();
+	
+	//Log user variables
 	private JLabel lblTitulo = new JLabel("Logged in as: ");
 	private JLabel lblTituloUser = new JLabel("");
-	private JLabel lblProjeto = new JLabel("Projetos");
+	
+	//Menu variables
 	private JMenuBar barraMenu = new JMenuBar();
 	private JMenu menuOpcao = new JMenu("Navegar");
 	private JMenuItem menuItem = new JMenuItem("Menu Partes");
 	private JMenuItem menuItem2 = new JMenuItem("Menu Projetos");
+	
+	//Menu display variables
+	private JLabel lblProjeto = new JLabel("Projetos");
 	private DefaultListModel<String> dlm2 = new DefaultListModel<String>();
 	private JList<String> listProjects = new JList<String>();
 	private JScrollPane jspProjects = new JScrollPane(listProjects);
+	
+	//Button variables
 	private JButton btnVer = new JButton("Ver");
 	private JButton btnAdicionar = new JButton("Adicionar");
 	private JButton btnMudar = new JButton("Alterar");
 	private JButton btnRemover = new JButton("Remover");
+	
+	//Preview variables
 	private JLabel lblPreview = new JLabel("Preview");
 	private JTextArea aPreview = new JTextArea();
-	private JScrollPane jspPreview = new JScrollPane(aPreview);
+	private JScrollPane jspPreview = new JScrollPane(aPreview, JScrollPane.VERTICAL_SCROLLBAR_ALWAYS, JScrollPane.HORIZONTAL_SCROLLBAR_NEVER);
 	
-	
-
+	//Constructor
+	@SuppressWarnings("static-access")
 	public ProjectMenu(String user) throws Exception {
-		//frame config
-		this.user = user;
+		//Frame configuration
 		fProjectMenu.setTitle("Menu Projetos");
 		fProjectMenu.setSize(1200,800);
 		fProjectMenu.setResizable(false);
 		fProjectMenu.setDefaultCloseOperation(fProjectMenu.EXIT_ON_CLOSE);
 		fProjectMenu.setLocationRelativeTo(null);
 	
+		//Menu configuration
 		fProjectMenu.setJMenuBar(barraMenu);
 		barraMenu.add(menuOpcao);
-		menuOpcao.add(menuItem); //a ser ativado
-		menuOpcao.add(menuItem2); //inativo por agora
+		menuOpcao.add(menuItem); 
+		menuOpcao.add(menuItem2); 
 		
 		menuItem.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
@@ -70,27 +80,29 @@ public class ProjectMenu {
 			}
 		});
 		
+		//Panel configuration
 		pn.setLayout(null);
 		pn.setBackground(Color.white);
 		fProjectMenu.add(pn);
 		
-		//titulo config
+		//Title configuration
 		lblTitulo.setBounds(320,40,300,65);
 		lblTitulo.setFont(new Font("Serif", Font.BOLD, 40));
 		pn.add(lblTitulo);
 		
+		//Title user configuration
 		lblTituloUser.setBounds(620,40,500,65);
 		lblTituloUser.setFont(new Font("Comic Sans", Font.BOLD, 40));
 		pn.add(lblTituloUser);
 		lblTituloUser.setText(user);
-		System.out.println("Current user(ProjMenu)" + user);
-		System.out.println("-----");
+		System.out.println("Current user (ProjMenu): " + user);
 		
-		//projects display config
+		//Project display configuration
 		lblProjeto.setBounds(150, 150, 200, 45);
 		lblProjeto.setFont(new Font("Serif", Font.ITALIC, 30));
 		pn.add(lblProjeto);
 		
+		//Display all projects of user configuration
 		jspProjects.setBounds(150,210,900,200);
 		listProjects.setFont(new Font("Serif", Font.BOLD, 20));
 		listProjects.setModel(dlm2);
@@ -100,12 +112,15 @@ public class ProjectMenu {
 
 		}
 		
-		//buttons config
+		//View button configuration
 		btnVer.setBounds(150,420,200,60);
 		btnVer.setFont(new Font("Serif", Font.PLAIN, 25));
 		pn.add(btnVer);
 		btnVer.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
+				if (listProjects.getSelectedIndex() == -1) {
+					return;
+				}
 				try {
 					new ProjectView(user, dlm2.getElementAt(listProjects.getSelectedIndex()).substring(5)).getJFrame().setVisible(true);
 				} catch (Exception e1) {
@@ -115,55 +130,68 @@ public class ProjectMenu {
 			}
 		});
 		
+		//Add button configuration
 		btnAdicionar.setBounds(380,420,200,60);
 		btnAdicionar.setFont(new Font("Serif", Font.PLAIN, 25));
 		pn.add(btnAdicionar);
 		
+		//Alter button configuration
 		btnMudar.setBounds(620,420,200,60);
 		btnMudar.setFont(new Font("Serif", Font.PLAIN, 25));
 		pn.add(btnMudar);
 		
+		//Remove button configuration
 		btnRemover.setBounds(850,420,200,60);
 		btnRemover.setFont(new Font("Serif", Font.PLAIN, 25));
 		pn.add(btnRemover);
 		
-		int confirm = 0;
 		btnRemover.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				int confirm = JOptionPane.showConfirmDialog(null, "Deletar projeto?", "Confimrar", JOptionPane.YES_NO_OPTION);
-				
-				if (confirm == JOptionPane.YES_OPTION) {
-					JOptionPane.showMessageDialog(null, "Projeto deletado");
-					try {
-						Project.deleteProjeto(dlm2.getElementAt(listProjects.getSelectedIndex()).substring(5));
-						System.out.println(dlm2.getElementAt(listProjects.getSelectedIndex()).substring(6));
-					} catch (Exception e1) {
-						e1.printStackTrace();
-					}
-					dlm2.removeAllElements();
-					try {
-						for (int i = 0; i < Project.getProjectSize(user); i++) {
-							dlm2.add(i, "NAME: " + Project.select(user, i, 0));
+				//Don't do anything if no project is selected
+				if (listProjects.getSelectedIndex() == -1){
+					return;
+				} else {
+					//Confirmation to remove
+					int confirm = JOptionPane.showConfirmDialog(null, "Deletar projeto?", "Confimrar", JOptionPane.YES_NO_OPTION);
+					if (confirm == JOptionPane.YES_OPTION) {
+						JOptionPane.showMessageDialog(null, "Projeto deletado");
+						try {
+							Project.deleteProjeto(dlm2.getElementAt(listProjects.getSelectedIndex()).substring(5));
+						} catch (Exception e1) {
+							e1.printStackTrace();
 						}
-					} catch (Exception e1) {
-						e1.printStackTrace();
+						
+						//Reload all elements 
+						dlm2.removeAllElements();
+						try {
+							for (int i = 0; i < Project.getProjectSize(user); i++) {
+								dlm2.add(i, "NAME: " + Project.select(user, i, 0));
+							}
+						} catch (Exception e1) {
+							e1.printStackTrace();
+						}
+					} else if (confirm == JOptionPane.NO_OPTION) {
+						JOptionPane.showMessageDialog(null, "Processo cancelado");
 					}
-				} else if (confirm == JOptionPane.NO_OPTION) {
-					JOptionPane.showMessageDialog(null, "Processo cancelado");
 				}
+
 			}
 		});
 		
-		//preview display config
+		//Preview display configuration
 		lblPreview.setBounds(150,500, 300,40);
 		lblPreview.setFont(new Font("Serif", Font.ITALIC, 30));
 		pn.add(lblPreview);
 		
+		//Project preview display
 		jspPreview.setBounds(150,550,900,150);
 		pn.add(jspPreview);
 
+		//Display preview of chosen project
 		listProjects.setSelectedIndex(0);
 		aPreview.setLineWrap(true);
+		aPreview.setEditable(false);
+		//Show on click
 		listProjects.addMouseListener(new MouseListener() {
 			public void mouseClicked(MouseEvent e) {
 				System.out.println(listProjects.getSelectedIndex());
@@ -186,7 +214,7 @@ public class ProjectMenu {
 
 	}
 
-	//métodos de accesso
+	//Class access functions
 	public JFrame fProjectMenu() {
 		return fProjectMenu;
 	}
